feat(nanoMOBILE): 100% native Nano Mobile Automation Engine, atomic perception & verified gestures
- Implement native Nano Mobile Automation Engine with EXECUTED != VERIFIED invariant
- Add NanoAtomicSnapshotter in Kotlin for simultaneous atomic visual/hierarchy captures
- Implement AdaptiveSwipeCalculator for dynamic container-bounded gestures
- Enforce NanoSensitiveDataPolicy for automatic credential & OTP redaction
- Expose standardized local MCP tool catalog with honest idempotency hints
- Add dual-step pruned operational memory in NanoTranscriptLedger
- Enforce strict < 300 lines limit across all subsystem files (100% SOLID)
- Validate 100% test suite (149/149 passed) and on-device hardware execution (Android 15)
- Update README and architecture documentation with Mermaid diagrams

+++ products/nanoRUNTIME/nanortime-ffi/src/lib.rs
@@ -766,7 +766,7 @@ impl NanoContext {
             .cached_tokens
             .len()
             .saturating_add(n_prompt)
-            .saturating_add(params.max_tokens as usize);
+            .saturating_add(params.max_tokens);
         if total_budget > self.context_size as usize {
             tracing::warn!(
                 "[NanoContext] contexto lleno (cached={} + prompt={} + max={} > ctx={}) — KV reset, prefill completo",
